+++ .gitignore
@@ -1,6 +1,5 @@
 # for a clean master branch
 dehb/examples/*
-dehb/utils/
 
 
 # files dumped at base directory
