+++ dehb/__init__.py
@@ -1,2 +1,3 @@
 from .optimizers import DE, AsyncDE
 from .optimizers import DEHB
+from .utils import SHBracketManager
