+++ dehb/utils/__init__.py
@@ -1,0 +1,1 @@
+from .bracket_manager import SHBracketManager
